feat(api): implement database migration system for all DBMS
Sets up a complete TypeORM migration infrastructure supporting all 5
databases (PostgreSQL, MySQL, MariaDB, MongoDB, SQLite), closing issue #19.

- Extract shared buildDataSourceOptions() from database.config.ts to
  avoid duplication between NestJS runtime config and TypeORM CLI
- Add standalone data-source.ts for the TypeORM CLI with proper dotenv
  loading order (required by @UnifiedId() decorator timing constraint)
- Add tsconfig.migration.json (commonjs override) for CLI compatibility
  with the project's nodenext module resolution
- Create per-DB migration directories: each DB type gets its own
  migrations since TypeORM generates DB-specific SQL
- Add initial schema migrations for all 5 databases (the user table)
- Add migration:generate/create/run/revert/show npm scripts
- Add migration tasks to api/Taskfile.yml and root Taskfile.yml
- Enable migrationsRun in production so app auto-runs pending migrations
  on startup; synchronize remains active in development
- Add api/MIGRATIONS.md with full developer documentation
