+++ .gitignore
@@ -12,3 +12,6 @@ mongodb/data
 .vscode/
 *.swp
 *.swo
+*.env
+*.env.local
+*.env.*.local
